+++ packages/client/test/unit/Influxdb.test.ts
@@ -1,12 +1,12 @@
 import {expect} from 'chai'
 import sinon from 'sinon'
 import {
-  InfluxDBClient,
   ClientOptions,
+  DEFAULT_ConnectionOptions,
+  InfluxDBClient,
   Transport,
   WriteOptions,
   WritePrecision,
-  DEFAULT_ConnectionOptions,
 } from '../../src'
 import type WriteApi from '../../src/WriteApi'
 import type QueryApi from '../../src/QueryApi'
@@ -971,4 +971,68 @@ at 'ClientOptions.database'
       expect(options).to.deep.equal({headers: {}, params: {}})
     })
   })
+  describe('call rest endpoints', () => {
+    it('should return server version success', async () => {
+      nock('http://test:8086').get('/ping').reply(
+        200,
+        {version: '3.0'},
+        {
+          'X-Influxdb-Version': '2.0',
+        }
+      )
+      const version = await getInfuxDbClient().getServerVersion()
+      expect(version).to.equal('2.0')
+    })
+    it('should return server version success', async () => {
+      nock('http://test:8086').get('/ping').reply(
+        200,
+        {version: '3.0'},
+        {
+          'x-influxdb-version': '2.0',
+        }
+      )
+      const version = await getInfuxDbClient().getServerVersion()
+      expect(version).to.equal('2.0')
+    })
+    it('should return server version success from body', async () => {
+      nock('http://test:8086')
+        .get('/ping')
+        .reply(200, {version: '3.0'}, {something: '2.0'})
+      const version = await getInfuxDbClient().getServerVersion()
+      expect(version).to.equal('3.0')
+    })
+    it('should return undefined server version', async () => {
+      nock('http://test:8086').get('/ping').reply(
+        200,
+        {notVersion: '3.0'},
+        {
+          something: '2.0',
+        }
+      )
+      const version = await getInfuxDbClient().getServerVersion()
+      expect(version).to.equal(undefined)
+    })
+    it('should throws exception', () => {
+      nock('http://test:8086').get('/ping').reply(
+        400,
+        {something: '3.0'},
+        {
+          'Some-header': '2.0',
+        }
+      )
+      getInfuxDbClient()
+        .getServerVersion()
+        .catch((reason) => {
+          expect(reason).to.equal('400')
+        })
+    })
+  })
 })
+
+function getInfuxDbClient() {
+  return new InfluxDBClient({
+    host: 'http://test:8086',
+    token: 'TEST_TOKEN',
+    database: 'TEST_DATABASE',
+  })
+}