+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Features
 
-1. [#588](https://github.com/InfluxCommunity/influxdb3-js/pull/588): Support grpc options available via [@grpc/grpc-js](https://github.com/grpc/grpc-node/blob/master/packages/grpc-js/README.md#supported-channel-options).
+1. [#585](https://github.com/InfluxCommunity/influxdb3-js/pull/585): Add function to get InfluxDB version.
+2. [#588](https://github.com/InfluxCommunity/influxdb3-js/pull/588): Support grpc options available via [@grpc/grpc-js](https://github.com/grpc/grpc-node/blob/master/packages/grpc-js/README.md#supported-channel-options).
    - New client option (`grpcOptions`) added. Accepts key-value pairs available via @grpc/grpc-js.
    - New queryOptions option (`grpcOptions`) added. Accepts key-value pairs available via @grpc/grpc-js.
    - Configurable also with the environment variable (`INFLUXDB_GRPC_OPTIONS`) which takes a comma separated list of key-value pairs.